Update all files to comply with ts rules, use module exports and imports

+++ System/Desk/DIUploadButton.ts
@@ -1,38 +1,33 @@
-/*
-** Class	: DIButton
-** 
-** This is simple button for the system
-** 
-** properties
-** 	-x				: x coordinate
-**	-y				: y coordinate
-**	-body			: Body of the view as HTML element
-**	-child			: Array of child views of this view
-**
-*/
+import { DeskEvent } from "../Secretary/DeskEvent";
+import { DIButton } from "./DIButton";
 
-class DIUploadButton extends DIButton {
-	constructor(text, className, idName) {
-		super(text, className, idName);
-		this.inputBody = document.createElement('input');
-		this.inputBody.setAttribute('type','file');
-		this.inputBody.style.display='none';
-		this.body.appendChild(this.inputBody);
-		this.buttonEvent = new DeskEvent(this.buttonBody, "click", function() {
-			this.inputBody.click();
-		}.bind(this));
-	}
-	
-	setButtonEvent(evt) {
-		if(this.event)
-			this.event.delete();
-		this.event = new DeskEvent(this.inputBody, "change", evt);
-	}
-	
-	delete() {
-		this.inputBody.remove();
-		this.buttonEvent.delete();
-		this.buttonEvent = null;
-		super.delete();
-	}
+/**
+ * This is simple button for the system
+ */
+export class DIUploadButton extends DIButton {
+    inputBody: HTMLInputElement;
+    buttonEvent: any;
+
+    constructor(text: string, className?: string, idName?: string) {
+        super(text, className, idName);
+        this.inputBody = document.createElement("input");
+        this.inputBody.setAttribute("type", "file");
+        this.inputBody.style.display = "none";
+        this.body.appendChild(this.inputBody);
+        this.buttonEvent = new DeskEvent(this.buttonBody, "click", () => {
+            this.inputBody.click();
+        });
+    }
+
+    setButtonEvent(evt: any) {
+        if (this.event) this.event.delete();
+        this.event = new DeskEvent(this.inputBody, "change", evt);
+    }
+
+    delete() {
+        this.inputBody.remove();
+        this.buttonEvent.delete();
+        this.buttonEvent = null;
+        super.delete();
+    }
 }
